Viewport Window!
The Viewport Window is now working. It can be docked and undocked like any other window. A bunch of work had to be done in order to debug the many problems with that, including making a fork of the imgui repository to make a small change in Vulkan implementation file, managing multiple GLFW focused/hovered windows, fixes in the Vulkan Barriers for proper stage resolution, and many more.

I am now going to work on the Hierarchy Window, and start adding a removing entities from the world!

List of Changes:
* YAML Serialization Wrapper, to fit with my serialization code. I will have to update the Components to use this code as well.
* Viewport Window complete
  * Many bug fixes for the docking capabilities.
  * Lock Aspect ratio feature
  * Completely dockable/undockable.
* Basic Console Window
  * Just prints the logs, which aren't enabled release mode at the moment. I would have to have Editor Configurations that allow game logs.
* The Editor settings are loaded from YAML, including things like the camera speed, sensitivity.
  * Basic overlay for changing those settings, that are attached to the viewport.

+++ Source/Engine/Nessie/Math/IVec2.inl
@@ -99,7 +99,7 @@ namespace nes
     template <IntegralType Type>
     float TIntVec2<Type>::LengthSqr() const
     {
-        return x * x + y * y;
+        return static_cast<float>(x * x + y * y);
     }
 
     template <IntegralType Type>
